Belgium BBAN check fix
The Belgian BBAN check has an issue when the checksum digits are exactly
"97". This is because the BBAN calculation is done using the "account
number" mod 97; normally, this would result in a checksum digit of "00",
but the standard in that case uses "97" instead.

Updated the checksum calculation to handle this case, plus updated the
related unit test.