Fix Polish IBAN specification
Polish IBANs have only one 8-digit bank code (often referred
in official documents as "bank branch number" or "branch
identifier") – it should not be split into three segments,
as was done in version 2024.01.1.

The changes in v2024.01.1 were based on a Wiki article that
was clearly incorrect. The official SWIFT document referenced
in the article (https://www.swift.com/resource/iban-registry-pdf)
always provided this information. The same information can be
found on the official website of the National Bank of Poland:
https://nbp.pl/pl/statystyka-i-raportowanie-finansowe/rynki-finansowe/iban-bic/

The Wiki article was recently updated when we came across it.

Although it's called a "branch identifier" or "bank branch number",
I believe in schwifty's terminology it's simply "bank_code."
(Because setting the bank_code length to 0 and the branch_code
length to 8 would be even more confusing)

Closes #259